watchdog-cam/requirements.txt
firebase-admin
opencv-python
google-generativeai
python-dotenv
requests
